refactor: teamId 기반 teamKey 추출
