+++ src/pages/team-dashboard-page.tsx
@@ -6,15 +6,27 @@ import MatchSchedule from '../components/Team/Match/match-schedule';
 import TeamPlayer from '../components/Team/Player/team-player';
 import TeamNews from '../components/common/TeamNews/team-news';
 
+import { useParams } from 'react-router-dom';
+import { TeamKey, teamKeyMap } from '../utils/team-key-map';
+
 const TeamDashboardPage = () => {
+  const { teamId } = useParams<{ teamId: string }>();
+
+  const teamKey =
+    teamId && teamKeyMap[teamId] ? (teamKeyMap[teamId] as TeamKey) : null;
+
+  if (!teamKey) {
+    return <div className="text-center">잘못된 팀 주소입니다.</div>;
+  }
+
   return (
     <div>
-      <TeamBanner teamKey="kia" />
+      <TeamBanner teamKey={teamKey} />
       <MatchSchedule />
       <MatchRecord />
       <TeamPlayer />
       <FanComment />
-      <Highlight teamKey="kia" />
+      <Highlight teamKey={teamKey} />
       <TeamNews />
     </div>
   );
