+++ src/utils/team-key-map.ts
@@ -1,0 +1,22 @@
+export type TeamKey = 'kia' | 'lg';
+//   | 'samsung'
+//   | 'doosan'
+//   | 'kt'
+//   | 'ssg'
+//   | 'lotte'
+//   | 'hanwha'
+//   | 'nc'
+//   | 'kiwoom';
+
+export const teamKeyMap: Record<string, TeamKey> = {
+  kiatigers: 'kia',
+  lgtwins: 'lg',
+  //   samsunglions: 'samsung',
+  //   doosanbears: 'doosan',
+  //   ktwiz: 'kt',
+  //   ssglanders: 'ssg',
+  //   lottegiants: 'lotte',
+  //   hanwhaeagles: 'hanwha',
+  //   ncdinos: 'nc',
+  //   kiwoomheroes: 'kiwoom',
+};
